+++ modules/chat_tools.py
@@ -14,6 +14,8 @@ from langchain.tools import StructuredTool
 import holoviews as hv
 import hvplot.xarray
 
+from modules.image_processing import s2_contrast_stretch, s2_image_to_uint8
+
 # import folium
 # from folium.raster_layers import TileLayer
 
@@ -24,34 +26,6 @@ import numpy as np
 import pandas as pd
 from datetime import datetime
 from holoviews.operation.datashader import rasterize
-
-
-# TODO: move these back to some module
-def s2_contrast_stretch(in_data):
-    """
-    Image enhancement: Contrast stretching.
-    """
-
-    p2, p98 = np.percentile(in_data.values.ravel(), (2.5, 97.5))
-    print(f"scaling to range {p2} : {p98}")
-    in_data.values = exposure.rescale_intensity(in_data, in_range=(p2, p98))
-
-    return in_data
-
-
-def s2_image_to_uint8(in_data):
-    """
-    A function that converts image DN to Reflectance (0, 1) and
-    then rescale to uint8 (0-255).
-    https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l1c/
-    """
-
-    # Convert to reflectance and uint8 (range: 0-255)
-    quant_value = 1e4
-    out_data = (in_data / quant_value * 255).astype("uint8")
-    out_data = out_data.clip(0, 255)
-
-    return out_data
 
 
 class MapManager(param.Parameterized):
@@ -115,6 +89,7 @@ class MapManager(param.Parameterized):
             "count": result.matched(),
         }
 
+    # TODO: switch to holoviews? or ok for demo?
     def view_footprints(
         self,
         # latitude: float,
@@ -126,7 +101,7 @@ class MapManager(param.Parameterized):
             self.gdf.loc[:, ["geometry"]]
             .set_crs("epsg:4326")
             .explore(tiles="CartoDB positron")
-        )  # TODO: basemap updates
+        )  # TODO: could use basemap, have optional tooltips, etc.
 
         chat_box.append({"SatGPT": pn.pane.plot.Folium(m, height=400)})
         return "Map is loaded to chat. Return nothing but a text confirmation to let the user know."
@@ -182,7 +157,7 @@ class MapManager(param.Parameterized):
             "Datacube is loaded to chat. Return nothing other than 'Done!' to the user."
         )
 
-    # TODO: make private or move to module
+    # TODO: replace with modules.image_plots
     def s2_hv_plot(
         self,
         items,
