+++ modules/image_processing.py
@@ -14,7 +14,20 @@ def s2_dn_to_reflectance(in_data):
 
     return in_data
 
+def s2_image_to_uint8(in_data):
+    """
+    A function that converts image DN to Reflectance (0, 1) and
+    then rescale to uint8 (0-255).
+    https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l1c/
+    """
 
+    # Convert to reflectance and uint8 (range: 0-255)
+    quant_value = 1e4
+    out_data = (in_data / quant_value * 255).astype("uint8")
+    out_data = out_data.clip(0, 255)
+
+    return out_data
+    
 def s2_contrast_stretch(in_data):
     """
     Image enhancement: Contrast stretching.
